+++ src/@locale/demo/metrics/index.js
@@ -2,6 +2,7 @@ import barList      from './barList' ;
 import categoryBar  from './categoryBar' ;
 import delta        from './delta' ;
 import metricLegend from './metricLegend' ;
+import metricScale  from './metricScale' ;
 import sparkline    from './sparkline' ;
 import tracker      from './tracker' ;
 
@@ -11,6 +12,7 @@ const metrics =
     categoryBar ,
     delta ,
     metricLegend ,
+    metricScale ,
     sparkline ,
     tracker ,
 } ;